Dao callback classes moved to dao.utils package; DB fields names must be constants, placed into dao.utils.UserRowMapper; Created table "roles" for storing roles values instead of user_roles.names (column will be removed); Created UserResultExtractor callback for retrieving users with roles by single query;

+++ src/main/java/com/epam/trn/dao/UserDao.java
@@ -26,7 +26,7 @@ public interface UserDao {
 
 	public void insertUserRoles(User user);
 	
-	public List<UserRole> getUserRoles(Integer userId);
+	public List<UserRole> getUserRoles(Long userId);
 
 	public int getTotal();
 }
